+++ Neuro/ActivationFunctions/AbsoluteReLU.cs
@@ -2,16 +2,16 @@
 {
     public class AbsoluteReLU : IActivationFunction
     {
-        public float Alpha { get; set; }
-        public float MinRange { get; set; } = -1;
-        public float MaxRange { get; set; } = 1;
+        public double Alpha { get; set; }
+        public double MinRange { get; set; } = -1;
+        public double MaxRange { get; set; } = 1;
         
-        public float Activation(float x)
+        public double Activation(double x)
         {
             return x >= 0 ? x : -x;
         }
 
-        public float Derivative(float x)
+        public double Derivative(double x)
         {
             return x >= 0 ? 1 : -1;
         }
