+++ Neuro/ActivationFunctions/ELU.cs
@@ -4,18 +4,18 @@ namespace Neuro.ActivationFunctions
 {
     public class ELU : IActivationFunction
     {
-        public float Alpha { get; set; }
-        public float MinRange { get; set; } = 0;
-        public float MaxRange { get; set; } = 1;
+        public double Alpha { get; set; }
+        public double MinRange { get; set; } = 0;
+        public double MaxRange { get; set; } = 1;
         
-        public float Activation(float x)
+        public double Activation(double x)
         {
-            return x >= 0 ? x : (float)Math.Exp(x) - 1;;
+            return x >= 0 ? x : (double)Math.Exp(x) - 1;;
         }
 
-        public float Derivative(float x)
+        public double Derivative(double x)
         {
-            return x >= 0 ? 1 : (float)Math.Exp(x);
+            return x >= 0 ? 1 : (double)Math.Exp(x);
         }
     }
 }
