+++ Neuro/ActivationFunctions/None.cs
@@ -4,16 +4,16 @@ namespace Neuro.ActivationFunctions
 {
     public class None : IActivationFunction
     {
-        public float Alpha { get; set; }
-        public float MinRange { get; set; } = -1;
-        public float MaxRange { get; set; } = 1;
+        public double Alpha { get; set; }
+        public double MinRange { get; set; } = -1;
+        public double MaxRange { get; set; } = 1;
         
-        public float Activation(float x)
+        public double Activation(double x)
         {
             return x;
         }
 
-        public float Derivative(float x)
+        public double Derivative(double x)
         {
             return 1;
         }
