+++ Neuro/Layers/FullyConnectedLayer.cs
@@ -12,7 +12,7 @@ namespace Neuro.Layers
         public LayerType Type { get; set; } = LayerType.FullyConnected;
         public ActivationType ActivationFunctionType { get; }
         public FullyConnectedNeuron[] Neurons { get; }
-        public float[] Outputs { get; private set; }
+        public double[] Outputs { get; private set; }
         public int NeuronsCount => Neurons.Length;
         public FullyConnectedNeuron this[int index] => Neurons[index];
         public IActivationFunction Function { get; }
@@ -23,7 +23,7 @@ namespace Neuro.Layers
             Function = activationType.Get();
             neuronsCount = Math.Max(1, neuronsCount);
             Neurons = new FullyConnectedNeuron[neuronsCount];
-            Outputs = new float[neuronsCount];
+            Outputs = new double[neuronsCount];
 
         }
         
@@ -37,7 +37,7 @@ namespace Neuro.Layers
                 Neurons[i] = new FullyConnectedNeuron(inputsCount, activationFunction);
             }
 
-            Outputs = new float[neuronsCount];
+            Outputs = new double[neuronsCount];
 
         }
 
@@ -57,9 +57,9 @@ namespace Neuro.Layers
             }
         }
 
-        public float[] Compute(float[] inputs)
+        public double[] Compute(double[] inputs)
         {
-            var outputs = new float[Outputs.Length];
+            var outputs = new double[Outputs.Length];
             
             Parallel.For(0, Neurons.Length, i =>
             {
